requirements.txt
SQLAlchemy~=2.0.25
python-dotenv~=1.0.0
fastapi~=0.108.0
pydantic~=2.5.3
requests~=2.31.0
uvicorn~=0.25.0